feat: refactor #1
