+++ apps/web/core/blocks/data/use-pagination.ts
@@ -1,0 +1,30 @@
+import * as React from 'react';
+
+export function usePagination() {
+  const [pageNumber, setPageNumber] = React.useState(0);
+
+  const setPage = React.useCallback(
+    (page: number | 'next' | 'previous') => {
+      switch (page) {
+        case 'next':
+          setPageNumber(prev => prev + 1);
+          break;
+        case 'previous': {
+          setPageNumber(prev => {
+            if (prev - 1 < 0) return 0;
+            return prev - 1;
+          });
+          break;
+        }
+        default:
+          setPageNumber(page);
+      }
+    },
+    [setPageNumber]
+  );
+
+  return {
+    pageNumber,
+    setPage,
+  };
+}
