+++ apps/web/core/blocks/data/use-relation-block.ts
@@ -1,0 +1,9 @@
+import { useTableBlockInstance } from '~/core/state/table-block-store';
+
+export function useRelationsQueryBlock() {
+  const { entityId, spaceId, relationId } = useTableBlockInstance();
+
+  return {
+    data: undefined,
+  };
+}
