+++ apps/web/core/io/subgraph/errors.ts
@@ -1,8 +1,8 @@
-export class HttpError {
+export class HttpError extends Error {
   readonly _tag = 'HttpError';
 }
 
-export class JsonParseError {
+export class JsonParseError extends Error {
   readonly _tag = 'JsonParseError';
 }
